push/requirements.txt
Flask
Flask-SQLAlchemy
requests
kafka-python
psycopg2-binary
python-dotenv